# Streaming PromQL engine: binary arithmetic operations with one-to-one matching