--- a/pkg/streamingpromql/compat/errors.go
+++ b/pkg/streamingpromql/compat/errors.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package streamingpromql
+package compat
 
 import (
 	"errors"