… CHANGELOG.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operator/aggregation_test.go
… pkg/streamingpromql/operator/binary_operation_test.go
… pkg/streamingpromql/operator/binary_operator.go
… pkg/streamingpromql/operator/operator.go
… pkg/streamingpromql/operator/operator_test.go
… pkg/streamingpromql/operator/pool.go
… pkg/streamingpromql/operator/ring_buffer.go
… pkg/streamingpromql/operator/ring_buffer_test.go
… pkg/streamingpromql/query.go
… pkg/streamingpromql/testdata/ours/binary_operators.test
… pkg/streamingpromql/testdata/upstream/operators.test
… pkg/streamingpromql/testdata/upstream/operators.test.disabled
… pkg/streamingpromql/testing.go
… tools/benchmark-query-engine/main.go