#### What this PR does

This PR adds support for binary arithmetic operations with one-to-one matching to the streaming PromQL engine.

This means it now supports expressions like the following:

* `metric_a + metric_b`
* `metric_a - on (env) metric_b`
* `metric_a * ignoring (env) metric_b`

It also supports binary operations with the other supported expression types, for example `rate(metric_a[1m]) + rate(metric_b[1m])` and `sum(metric_a + metric_b)`.

The `+`, `-`, `*`, `/`, `%`, `^` and `atan2` operations are supported. 

Comparison operations like `==` and `>`, and logical operations like `or` and `unless` are not supported. One-to-many and many-to-one matching with `group_left` or `group_right` are also not supported. These features will be added in future PRs.

---

##### Performance

In general, the streaming engine is substantially faster than Prometheus' engine, particularly for queries that select many series or evaluate many steps - around 60-70% faster in many cases. Similarly, the streaming engine has far lower peak memory utilisation in all but one scenario, with 60-70% lower memory utilisation for queries that select many series and many steps.

The only scenario where peak memory utilisation is not lower with the streaming engine is for queries where one side of the binary operator has more series than the other (`a_100{l=~"[13579]."} - b_100` and `a_2000{l=~"1..."} - b_2000` below). I haven't yet been able to get to the bottom of this, and will follow up with a separate PR to fix the issue rather than wait and make this PR even bigger.

<details>
<summary>Detailed benchmark comparison with Prometheus' engine</summary>

First section below shows latency, second section shows peak memory utilisation. 

`a_X` and `b_X` in the queries below are metrics with X series.

```
                                                                       │      standard      │              streaming              │
                                                                       │       sec/op       │    sec/op     vs base               │
Query/a_1_-_b_1,_instant_query                                                 225.9µ ±  1%   250.8µ ±  6%  +11.00% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps                                    260.3µ ±  0%   260.6µ ± 14%        ~ (p=1.000 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps                                   618.3µ ±  2%   365.8µ ±  2%  -40.83% (p=0.002 n=6)
Query/a_100_-_b_100,_instant_query                                             1.182m ±  1%   1.159m ± 16%        ~ (p=0.065 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps                                4.437m ±  1%   2.039m ±  2%  -54.04% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps                               33.51m ±  0%   10.20m ±  1%  -69.56% (p=0.002 n=6)
Query/a_2000_-_b_2000,_instant_query                                           15.92m ±  1%   15.40m ±  1%   -3.28% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps                              82.01m ±  1%   32.78m ±  1%  -60.02% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps                             735.9m ±  1%   185.6m ±  0%  -74.78% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps                                  3.298m ±  1%   1.279m ±  1%  -61.22% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps                             322.21m ±  0%   86.60m ±  1%  -73.12% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps                             7.571 ±  0%    1.728 ±  6%  -77.17% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query                              1.114m ±  1%   1.092m ±  3%   -1.98% (p=0.041 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps                 3.182m ±  1%   1.773m ±  1%  -44.27% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps               21.739m ±  0%   8.026m ±  3%  -63.08% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query                               12.093m ±  1%   9.258m ±  1%  -23.44% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps                   52.50m ±  1%   18.50m ±  2%  -64.76% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps                  470.5m ±  1%   103.1m ±  1%  -78.09% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query                             232.9µ ±  1%   240.4µ ±  2%   +3.22% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps                276.6µ ±  2%   254.3µ ±  3%   -8.08% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps               710.4µ ±  2%   360.5µ ±  3%  -49.25% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query                         1.150m ±  1%   1.115m ±  1%   -3.03% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps            5.453m ±  1%   2.302m ±  1%  -57.77% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps           43.07m ±  1%   11.64m ±  1%  -72.98% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query                       15.28m ±  1%   14.94m ±  2%   -2.26% (p=0.015 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps         102.03m ±  1%   37.98m ±  1%  -62.78% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps         928.4m ±  1%   216.2m ±  3%  -76.71% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_instant_query                                            225.9µ ±  1%   233.0µ ± 18%   +3.17% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_range_query_with_100_steps                               266.8µ ±  2%   263.6µ ±  9%        ~ (p=0.310 n=6)
Query/sum(a_1_+_b_1),_range_query_with_1000_steps                              650.4µ ±  2%   351.5µ ±  3%  -45.95% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_instant_query                                        1.189m ±  1%   1.165m ±  2%   -2.04% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_range_query_with_100_steps                           4.504m ±  1%   2.158m ± 17%  -52.07% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_range_query_with_1000_steps                          34.07m ±  1%   10.51m ±  1%  -69.16% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_instant_query                                      15.94m ±  1%   15.51m ±  5%        ~ (p=0.310 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_100_steps                         83.55m ±  2%   33.44m ±  2%  -59.97% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_1000_steps                        764.1m ±  0%   191.0m ±  2%  -75.00% (p=0.002 n=6)

                                                                       │      standard      │               streaming                │
                                                                       │         B          │       B         vs base                │
Query/a_1_-_b_1,_instant_query                                                72.75Mi ±  1%    72.34Mi ±  2%         ~ (p=0.132 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps                                   71.61Mi ±  1%    71.88Mi ±  1%         ~ (p=0.485 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps                                  68.48Mi ±  1%    70.20Mi ±  1%    +2.51% (p=0.002 n=6)
Query/a_100_-_b_100,_instant_query                                            66.76Mi ±  1%    66.43Mi ±  2%         ~ (p=0.420 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps                               67.09Mi ±  1%    66.52Mi ±  1%    -0.84% (p=0.006 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps                              73.14Mi ±  3%    69.22Mi ±  1%    -5.36% (p=0.002 n=6)
Query/a_2000_-_b_2000,_instant_query                                          69.03Mi ±  1%    69.47Mi ±  1%         ~ (p=0.619 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps                             89.50Mi ±  3%    78.26Mi ±  1%   -12.56% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps                            213.2Mi ±  0%    139.5Mi ±  1%   -34.56% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps                                 68.86Mi ±  1%    70.04Mi ±  2%    +1.71% (p=0.026 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps                             158.6Mi ±  5%    113.2Mi ±  5%   -28.65% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps                          1600.8Mi ± 11%    460.1Mi ±  3%   -71.26% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query                             67.02Mi ±  1%    66.43Mi ±  2%         ~ (p=0.420 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps                66.80Mi ±  1%    66.62Mi ±  1%         ~ (p=1.000 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps               70.56Mi ±  1%    68.10Mi ±  1%    -3.49% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query                               68.22Mi ±  3%   139.30Mi ±  3%  +104.20% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps                  81.85Mi ±  1%   111.61Mi ± 37%   +36.36% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps                 189.6Mi ±  6%    175.0Mi ± 21%    -7.72% (p=0.004 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query                            72.02Mi ±  1%    72.05Mi ±  1%         ~ (p=0.894 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps               71.26Mi ±  1%    71.98Mi ±  1%         ~ (p=0.132 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps              67.88Mi ±  1%    69.65Mi ±  1%    +2.60% (p=0.004 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query                        66.96Mi ±  1%    65.92Mi ±  1%    -1.55% (p=0.009 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps           67.11Mi ±  1%    66.75Mi ±  2%         ~ (p=0.143 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps          72.39Mi ±  3%    68.47Mi ±  2%    -5.42% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query                      69.16Mi ±  2%    68.61Mi ±  1%         ~ (p=0.058 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps         86.86Mi ±  4%    77.52Mi ±  2%   -10.75% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps        211.6Mi ±  0%    139.5Mi ±  2%   -34.11% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_instant_query                                           72.75Mi ±  1%    72.49Mi ±  1%         ~ (p=0.370 n=6)
Query/sum(a_1_+_b_1),_range_query_with_100_steps                              71.85Mi ±  1%    72.67Mi ±  1%    +1.14% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_range_query_with_1000_steps                             68.41Mi ±  1%    74.67Mi ±  1%    +9.15% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_instant_query                                       66.73Mi ±  1%    66.48Mi ±  1%         ~ (p=0.310 n=6)
Query/sum(a_100_+_b_100),_range_query_with_100_steps                          66.50Mi ±  1%    66.18Mi ±  1%         ~ (p=0.132 n=6)
Query/sum(a_100_+_b_100),_range_query_with_1000_steps                         73.99Mi ±  1%    65.91Mi ±  2%   -10.92% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_instant_query                                     69.02Mi ±  1%    68.75Mi ±  1%         ~ (p=0.071 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_100_steps                        89.06Mi ±  2%    69.47Mi ±  3%   -22.00% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_1000_steps                      213.24Mi ±  0%    78.56Mi ±  2%   -63.16% (p=0.002 n=6)
```

</details>

---

##### Fix for issue where peak memory utilisation is much higher than Prometheus' engine for queries that select many series

This PR also fixes the issue where latency and peak memory utilisation is much higher than Prometheus' engine for queries that select many series. Peak memory utilisation for these queries is now near identical in these cases, while the streaming engine is still significantly faster in all benchmarks. See the commit message on https://github.com/grafana/mimir/commit/825b4dcba678221a8525f1262d144bf7230d70cd for more explanation of the fix.

<details>
<summary>Detailed benchmark comparison with Prometheus' engine</summary>

First section below shows latency, second section shows peak memory utilisation.

`a_X` and `b_X` in the queries below are metrics with X series.

```
                                                                       │      standard      │              streaming              │
                                                                       │       sec/op       │    sec/op     vs base               │
Query/a_1,_instant_query                                                       127.9µ ±  5%   120.6µ ±  4%   -5.72% (p=0.004 n=6)
Query/a_1,_range_query_with_100_steps                                          141.8µ ± 15%   131.8µ ±  8%        ~ (p=0.132 n=6)
Query/a_1,_range_query_with_1000_steps                                         181.0µ ±  2%   171.4µ ±  1%   -5.33% (p=0.002 n=6)
Query/a_100,_instant_query                                                     577.7µ ±  7%   558.4µ ±  3%   -3.35% (p=0.002 n=6)
Query/a_100,_range_query_with_100_steps                                        1.060m ±  5%   1.039m ±  5%        ~ (p=0.240 n=6)
Query/a_100,_range_query_with_1000_steps                                       5.276m ±  1%   5.191m ±  2%   -1.61% (p=0.015 n=6)
Query/a_2000,_instant_query                                                    7.207m ±  0%   7.015m ±  1%   -2.65% (p=0.002 n=6)
Query/a_2000,_range_query_with_100_steps                                       15.88m ±  1%   15.39m ±  1%   -3.08% (p=0.002 n=6)
Query/a_2000,_range_query_with_1000_steps                                      94.69m ±  1%   90.68m ±  0%   -4.24% (p=0.002 n=6)
Query/rate(a_1[1m]),_instant_query                                             123.1µ ±  7%   112.7µ ±  3%   -8.43% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                                134.4µ ±  4%   120.3µ ±  4%  -10.48% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                               236.3µ ±  1%   178.0µ ±  1%  -24.69% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                           570.8µ ±  2%   557.4µ ±  4%   -2.34% (p=0.004 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                              1.606m ±  2%   1.153m ±  0%  -28.20% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                            10.467m ±  2%   5.811m ±  1%  -44.48% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                          7.007m ±  0%   6.811m ±  1%   -2.80% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                             26.64m ±  1%   17.95m ±  2%  -32.62% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                            195.2m ±  2%   104.3m ±  0%  -46.57% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                             1174.8µ ±  2%   744.3µ ±  1%  -36.64% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                            97.00m ±  1%   52.99m ±  6%  -45.38% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps                            2.004 ±  2%    1.047 ±  5%  -47.75% (p=0.002 n=6)
Query/rate(a_1[1d]),_instant_query                                             745.6µ ±  2%   553.4µ ±  7%  -25.77% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps                               1302.3µ ±  1%   988.9µ ±  1%  -24.07% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps                               6.227m ±  1%   4.860m ±  1%  -21.94% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query                                           43.68m ±  0%   33.71m ±  0%  -22.83% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps                              98.27m ±  1%   76.75m ±  1%  -21.90% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps                             587.2m ±  1%   461.3m ±  0%  -21.44% (p=0.002 n=6)
Query/rate(a_2000[1d]),_instant_query                                          813.1m ±  1%   618.5m ±  0%  -23.93% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps                              1.914 ±  1%    1.482 ±  8%  -22.58% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps                            11.671 ±  1%    9.149 ±  0%  -21.61% (p=0.002 n=6)

                                                                       │      standard      │               streaming                │
                                                                       │         B          │       B         vs base                │
Query/a_1,_instant_query                                                      72.84Mi ±  1%    72.25Mi ±  1%    -0.82% (p=0.015 n=6)
Query/a_1,_range_query_with_100_steps                                         72.70Mi ±  1%    71.51Mi ±  1%    -1.64% (p=0.004 n=6)
Query/a_1,_range_query_with_1000_steps                                        72.05Mi ±  2%    70.04Mi ±  1%    -2.80% (p=0.002 n=6)
Query/a_100,_instant_query                                                    66.33Mi ±  2%    65.80Mi ±  1%         ~ (p=0.368 n=6)
Query/a_100,_range_query_with_100_steps                                       67.19Mi ±  1%    65.77Mi ±  1%    -2.12% (p=0.004 n=6)
Query/a_100,_range_query_with_1000_steps                                      68.54Mi ±  1%    68.41Mi ±  1%         ~ (p=0.394 n=6)
Query/a_2000,_instant_query                                                   68.42Mi ±  3%    68.45Mi ±  0%         ~ (p=1.000 n=6)
Query/a_2000,_range_query_with_100_steps                                      74.73Mi ±  1%    76.11Mi ±  1%    +1.84% (p=0.002 n=6)
Query/a_2000,_range_query_with_1000_steps                                     134.5Mi ±  1%    136.3Mi ±  9%         ~ (p=0.509 n=6)
Query/rate(a_1[1m]),_instant_query                                            71.85Mi ±  1%    71.84Mi ±  1%         ~ (p=0.818 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                               72.88Mi ±  2%    71.96Mi ±  1%         ~ (p=0.065 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                              70.88Mi ±  2%    69.79Mi ±  1%         ~ (p=0.067 n=6)
Query/rate(a_100[1m]),_instant_query                                          66.68Mi ±  1%    65.83Mi ±  0%    -1.28% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                             66.08Mi ±  1%    65.91Mi ±  2%         ~ (p=0.699 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                            67.52Mi ±  1%    68.53Mi ±  1%    +1.49% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                         69.01Mi ±  1%    68.42Mi ±  2%    -0.85% (p=0.015 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                            73.89Mi ±  2%    75.41Mi ±  3%    +2.06% (p=0.006 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                           131.7Mi ±  1%    132.8Mi ±  2%         ~ (p=0.292 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                             67.65Mi ±  1%    68.72Mi ±  1%    +1.58% (p=0.004 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                           103.6Mi ±  1%    112.3Mi ±  4%    +8.41% (p=0.002 n=6)
…